requirements.txt
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
fastapi
jinja2
python-multipart
langchain-chroma
faiss-cpu